requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
aioredis
confluent-kafka
mysqlclient
bcrypt
python-jose[cryptography]
